[Refactor] 서비스클래스에 인터페이스 도입#96
