+++ src/main/java/me/snaptime/snap/controller/SnapPagingController.java
@@ -6,7 +6,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import me.snaptime.common.CommonResponseDto;
 import me.snaptime.snap.dto.res.FindSnapPagingResDto;
-import me.snaptime.snap.service.impl.SnapPagingServiceImpl;
+import me.snaptime.snap.service.SnapPagingService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -20,7 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
 @Tag(name = "[Social] Snap API")
 public class SnapPagingController {
 
-    private final SnapPagingServiceImpl snapPagingService;
+    private final SnapPagingService snapPagingService;
 
     @Operation(summary = "Snap 조회", description = "커뮤니티에서 Snap을 10개씩 페이징조회합니다.")
     @Parameter(name = "pageNum", description = "Snap페이지 번호를 보내주세요")
